requirements.txt
Django==1.4
Pillow==2.5.3
wsgiref==0.1.2
xlrd==0.9.3
django-registration